requirements.txt
streamlit
snowflake-snowpark-python[pandas]
ipyvizzu
ipyvizzu-story
streamlit-extras
iso3166
openpyxl
